% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewSecret.R
\name{addNewSecret}
\alias{addNewSecret}
\title{Addition of key-value pairs.
If 'value' is not provided, then prompt for it.}
\usage{
addNewSecret(name, value = character(0), users, vault)
}
\arguments{
\item{name}{A string containing the name, or key, of the secret to be added.}

\item{value}{A string containing the value of the secret to be added.}

\item{users}{A list of current users who will have access to the secret.}

\item{vault}{A string containing the location of the vault.}
}
\value{
The value stored.
}
\description{
Addition of key-value pairs.
If 'value' is not provided, then prompt for it.
}
\examples{
addNewSecret( "remove_me_1234", value="junk",
    users=secret::list_users( get_secret_vault() )[1],
    vault=get_secret_vault() )
secret::delete_secret(name="remove_me_1234",vault=get_secret_vault() )
}
