% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSecretsWithSubstring.R
\name{deleteSecretsWithSubstring}
\alias{deleteSecretsWithSubstring}
\title{Clear any key-value pair where the pair contains the substring, ss.}
\usage{
deleteSecretsWithSubstring(name, vault)
}
\arguments{
\item{name}{A string containing the substring to be matching against all keys.}

\item{vault}{A string containing location of the vault.}
}
\value{
The vault
}
\description{
Clear any key-value pair where the pair contains the substring, ss.
}
\examples{
addNewSecret( "remove_me_1234", value="junk",
    users=secret::list_users( get_secret_vault() )[1],
    vault=get_secret_vault() )
clearSecretsWithSubstring(ss="remove_me_1234",vault=get_secret_vault() )
}
