% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret_vault.R
\name{get_secret_vault}
\alias{get_secret_vault}
\title{Create a data iterator for MEF files
Assumes that the user has already created an SSH key.}
\usage{
get_secret_vault()
}
\value{
A link to a vault. Creates a new one if none exist.
}
\description{
Create a data iterator for MEF files
Assumes that the user has already created an SSH key.
}
\examples{
vault <- get_secret_vault()
}
