% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_here.R
\name{init_here}
\alias{init_here}
\title{Override the tree search to set the root directory for here().}
\usage{
init_here()
}
\description{
Override the tree search to set the root directory for here().
}
\examples{
init_here()
}
